src/APIs/arctis_nova_api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arctis-nova-api"
version = "0.1.0"
description = "Python API for SteelSeries Arctis Nova Pro Wireless (Sonar, GameSense OLED, base station)"
readme = "../../../README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
]

[project.optional-dependencies]
usb = ["hidapi>=0.14.0"]
test = ["pytest>=8.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
